Improve timer reliability by waking background
The "alarms" API has an "arbitrary" delay as stated by Firefox and
Chrome docs and can't be trusted to be precise for the timer. Instead
let's use the alarms api to wake the background script earlier than the
timer and use setTimeout for the actual timer expiration. Right now I am
waking 25 seconds early since Chrome explicitly states service workers
are killed after 30 seconds of inactivity.

TODO: The badge alarms could use this same treatment, but it's a lower
priority.